+++ CHANGES.md
@@ -6,6 +6,7 @@
 * Add `reduce` and `reverse` method to ranges (#114)
 * Fix `orbit.global_velocity` to use correct frame of reference (#116)
 * Fix internal type comparison/generation of structs (#118)
+* Add `.sort_by` and `.sort_with` to array
 
 ## 0.5.2.2 -> 0.5.2.4